pyproject.toml
[tool.poetry]
name = "sanic-playground"
version = "0.1.0"
description = ""
authors = ["Monu Kumar <kmonu9581@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sanic = "^23.12.1"
tortoise-orm = "^0.21.3"
setuptools = "^70.1.1"
asyncpg = "^0.29.0"
redis = "^5.0.6"
pre-commit = "^3.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
